[NFC][wasm-reduce] Move Reducer to .h and .cpp files
wasm-split's Reducer class is very large and does many different things.
To start getting a handle on it, separate out the declarations and
definitions of its methods into new .h and .cpp files.
