+++ src/tools/wasm-reduce/CMakeLists.txt
@@ -1,6 +1,7 @@
 FILE(GLOB wasm_reduce_HEADERS *h)
 set(wasm_reduce_SOURCES
   wasm-reduce.cpp
+  reducer.cpp
   ${wasm_reduce_HEADERS}
 )
 binaryen_add_executable(wasm-reduce "${wasm_reduce_SOURCES}")
